--- axum-debug-macros/CHANGELOG.md
@@ -1,13 +1,0 @@
-# Changelog
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-# Unreleased
-
-- **breaking:** Removed `debug_router` macro.
-
-# 0.1.0 (6. October 2021)
-
-- Initial release.
