+++ axum-debug/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -9,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **breaking:** Removed `debug_router` macro.
 - **breaking:** Removed `check_service` function.
 - **breaking:** Removed `debug_service` function.
+- **breaking:** No longer re-exports `axum`.
 
 # 0.1.0 (6. October 2021)
